st/nine: specify WINAPI only for i386 and amd64
This is cherry-picked from commit 2a529a8ac810b3b40e7a9a071d9544b168928b94 from the mesa repository.

Currently mesa fails building with the x32 abi as ms_abi is not defined
in such a case.

The patch uses ms_abi only for amd64 targets and stdcall only for i386
targets to be sure that those are defined.

This patch additionally checks for __GNUC__ to guarantee that
__attribute__ is available.